Move cmd logic from `AppInitFlow` to `AppInitCmd`.
