+++ examples/app_cycle/src/cmds/app_init_cmd.rs
@@ -1,13 +1,16 @@
-use std::path::PathBuf;
-
-use peace::rt_model::output::OutputWrite;
-use peace_item_specs::{file_download::FileDownloadParams, tar_x::TarXParams};
+use peace::{
+    cfg::{app_name, AppName, Profile},
+    cmd::ctx::CmdCtx,
+    resources::states::StatesSaved,
+    rt::cmds::{EnsureCmd, StatesDiscoverCmd},
+    rt_model::{output::OutputWrite, Workspace, WorkspaceSpec},
+};
 use semver::Version;
 use url::Url;
 
 use crate::{
-    flows::AppInitFlow,
-    model::{AppCycleError, RepoSlug, WebAppFileId},
+    flows::EnvDeployFlow,
+    model::{AppCycleError, RepoSlug},
 };
 
 /// Takes app init parameters and runs the [`AppInitFlow`].
@@ -25,58 +28,49 @@ impl AppInitCmd {
     /// * `url`: URL to override where to download the application from.
     pub async fn run<O>(
         output: &mut O,
-        slug: RepoSlug,
-        version: Version,
+        slug: &RepoSlug,
+        version: &Version,
         url: Option<Url>,
     ) -> Result<(), AppCycleError>
     where
         O: OutputWrite<AppCycleError>,
     {
-        let account = slug.account();
-        let repo_name = slug.repo_name();
-        let web_app_download_dir = PathBuf::from_iter([account, repo_name, &format!("{version}")]);
+        let (web_app_file_download_params, web_app_tar_x_params) =
+            EnvDeployFlow::params(slug, version, url)?;
+        let workspace = Workspace::new(
+            app_name!(),
+            #[cfg(not(target_arch = "wasm32"))]
+            WorkspaceSpec::WorkingDir,
+            #[cfg(target_arch = "wasm32")]
+            WorkspaceSpec::SessionStorage,
+        )?;
+        let flow = EnvDeployFlow::flow().await?;
 
-        let web_app_file_download_params = {
-            #[cfg(target_family = "windows")]
-            let file_ext = "zip";
-            #[cfg(any(target_family = "unix", target_family = "wasm"))]
-            let file_ext = "tar";
-            // windows:
-            // https://github.com/azriel91/web_app/releases/download/0.1.0/web_app.zip
-            //
-            // linux:
-            // https://github.com/azriel91/web_app/releases/download/0.1.0/web_app.tar
-            let src = {
-                match url {
-                    Some(url) => url,
-                    None => {
-                        let url_candidate = format!(
-                            "https://github.com/{account}/{repo_name}/releases/download/{version}/{repo_name}.{file_ext}"
-                        );
-                        Url::parse(&url_candidate).map_err(|error| {
-                            AppCycleError::AppCycleUrlBuild {
-                                url_candidate,
-                                error,
-                            }
-                        })?
-                    }
-                }
-            };
-            let dest = web_app_download_dir.join(format!("{repo_name}.{file_ext}"));
-            FileDownloadParams::new(
-                src,
-                dest,
-                #[cfg(target_arch = "wasm32")]
-                peace_item_specs::file_download::StorageForm::Base64,
-            )
-        };
-        let web_app_tar_x_params = {
-            let tar_path = web_app_file_download_params.dest().to_path_buf();
-            let dest = web_app_download_dir.join("extracted");
+        let mut cmd_ctx = {
+            let cmd_ctx_builder =
+                CmdCtx::builder_single_profile_single_flow::<AppCycleError, _>(output, &workspace);
+            crate::cmds::ws_and_profile_params_augment!(cmd_ctx_builder);
 
-            TarXParams::<WebAppFileId>::new(tar_path, dest)
+            cmd_ctx_builder
+                .with_profile(Profile::workspace_init())
+                .with_flow(&flow)
+                .with_workspace_param_value(
+                    String::from("web_app_file_download_params"),
+                    Some(web_app_file_download_params),
+                )
+                .with_workspace_param_value(
+                    String::from("web_app_tar_x_params"),
+                    Some(web_app_tar_x_params),
+                )
+                .await?
         };
 
-        AppInitFlow::run(output, web_app_file_download_params, web_app_tar_x_params).await
+        let (states_current, _states_desired) = StatesDiscoverCmd::exec(&mut cmd_ctx).await?;
+        let states_saved = StatesSaved::from(states_current);
+
+        let states_ensured = EnsureCmd::exec(&mut cmd_ctx, &states_saved).await?;
+        cmd_ctx.output_mut().present(&states_ensured).await?;
+
+        Ok(())
     }
 }
